added forced reconfigurations to the list
Dynamic cluster membership will go via the paxos engine requiring a quorum of nodes to be up and reachable to change the configuration. If a majority of nodes go permanently offline then an administrator needs a way to force an update to the configuration to add new members to restore the number of nodes to the desired number of replicas.